move DowngradeParameterTypeWideningRector to Class_ namespace
