--- rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Fixture/keep_interface_downgraded.php.inc
@@ -1,12 +1,0 @@
-<?php
-
-namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
-
-use Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source\SomeContainerInterface;
-
-class KeepInterfaceDonwgraded implements SomeContainerInterface
-{
-    public function get($id)
-    {
-    }
-}
