--- rules-tests/DowngradePhp72/Rector/ClassMethod/DowngradeParameterTypeWideningRector/Source/AnyService.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source;
-
-final class AnyService
-{
-
-}
