+++ rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Fixture/keep_interface_downgraded.php.inc
@@ -1,0 +1,12 @@
+<?php
+
+namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
+
+use Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Source\SomeContainerInterface;
+
+class KeepInterfaceDonwgraded implements SomeContainerInterface
+{
+    public function get($id)
+    {
+    }
+}
