+++ rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Fixture/skip_command.php.inc
@@ -1,8 +1,8 @@
 <?php
 
-namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
 
-use Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Source\AnyService;
+use Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Source\AnyService;
 use Symfony\Component\Console\Command\Command;
 
 final class SkipCommand extends Command
