+++ rules-tests/DowngradePhp72/Rector/Class_/DowngradeParameterTypeWideningRector/Fixture/with_class_inheritance.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
 
 class AncestorClass
 {
@@ -22,7 +22,7 @@ class ChildClass extends AncestorClass
 -----
 <?php
 
-namespace Rector\Tests\DowngradePhp72\Rector\ClassMethod\DowngradeParameterTypeWideningRector\Fixture;
+namespace Rector\Tests\DowngradePhp72\Rector\Class_\DowngradeParameterTypeWideningRector\Fixture;
 
 class AncestorClass
 {
